+++ app/src/main/java/labs/zero_one/roundo/MapKit.kt
@@ -5,6 +5,8 @@ import android.content.res.Resources
 import android.location.Location
 import android.os.Bundle
 import com.minemap.minemapsdk.MinemapAccountManager
+import com.minemap.minemapsdk.annotations.Icon
+import com.minemap.minemapsdk.annotations.IconFactory
 import com.minemap.minemapsdk.annotations.Marker
 import com.minemap.minemapsdk.annotations.MarkerOptions
 import com.minemap.minemapsdk.camera.CameraPosition
@@ -22,12 +24,17 @@ import com.minemap.minemapsdk.maps.MineMap
  * - [markerList]
  * - [tempMarkerOptionsList]
  * - [isFollowing]
+ * - [markerIconList]
+ *
+ * ## 子类列表
+ * - [MarkerType]
  *
  * ## 方法列表
  * - [initMap]
  * - [moveTo]
  * - [add]
  * - [addMarkerAt]
+ * - [changeMarkerIconAt]
  * - [clearMarkers]
  *
  * @param [context] 环境
@@ -40,6 +47,7 @@ import com.minemap.minemapsdk.maps.MineMap
  * @property [tempMarkerOptionsList] 标记暂存列表
  * @property [markerList] 标记列表
  * @property [isFollowing] 中心是否跟随移动
+ * @property [markerIconList] 标记样式列表
  */
 class MapKit(private val context: Context) {
 
@@ -48,6 +56,24 @@ class MapKit(private val context: Context) {
     private var markerList: ArrayList<Marker> = ArrayList(0)
     private var tempMarkerOptionsList: ArrayList<MarkerOptions> = ArrayList(0)
     var isFollowing = true
+
+    /**
+     * 标记类型
+     *
+     * ## 列表
+     * - [Unchecked] 未打卡点，橙色
+     * - [Checked] 已打卡点，绿色
+     *
+     * @param [iconIndex] 图标在 [markerIconList] 中的序号
+     *
+     * @author lucka-me
+     * @since 0.1.10
+     */
+    enum class MarkerType(val iconIndex: Int) {
+        Unchecked(0),
+        Checked(1)
+    }
+    private lateinit var markerIconList: Array<Icon>
 
     /**
      * 初始化地图控制器
@@ -64,6 +90,10 @@ class MapKit(private val context: Context) {
             context,
             context.getString(R.string.minemap_token),
             "4810"
+        )
+        markerIconList = arrayOf(
+            IconFactory.getInstance(context).fromResource(R.mipmap.ic_marker_unchecked),
+            IconFactory.getInstance(context).fromResource(R.mipmap.ic_marker_checked)
         )
         mapView.onCreate(savedInstanceState)
         mapView.getMapAsync { newMap ->
@@ -86,7 +116,7 @@ class MapKit(private val context: Context) {
             mineMap.cameraPosition =
                 CameraPosition.Builder()
                     .target(LatLng(locationKit.lastLocation))
-                    .zoom(12.0)
+                    .zoom(14.0)
                     .build()
         }
     }
@@ -130,7 +160,7 @@ class MapKit(private val context: Context) {
     }
 
     /**
-     * 在指定位置添加标记
+     * 在指定位置添加默认标记
      *
      * @param [location] 目标位置
      *
@@ -139,6 +169,36 @@ class MapKit(private val context: Context) {
      */
     fun addMarkerAt(location: Location) {
         add(MarkerOptions().position(LatLng(location)))
+    }
+
+    /**
+     * 在指定位置添加指定标记
+     *
+     * @param [location] 目标位置
+     * @param [type] 标记类型
+     *
+     * @author lucka-me
+     * @since 0.1.10
+     */
+    fun addMarkerAt(location: Location, type: MarkerType) {
+        add(MarkerOptions().position(LatLng(location)).icon(markerIconList[type.iconIndex]))
+    }
+
+    /**
+     * 修改指定标记的样式
+     *
+     * @param [index] 标记在列表中的序号
+     * @param [type] 目的类型
+     *
+     * @author lucka-me
+     * @since 0.1.10
+     */
+    fun changeMarkerIconAt(index: Int, type: MarkerType) {
+        if (!isMapInitialized && index < tempMarkerOptionsList.size) {
+            tempMarkerOptionsList[index].icon = markerIconList[type.iconIndex]
+        } else if (isMapInitialized && index < markerList.size) {
+            markerList[index].icon = markerIconList[type.iconIndex]
+        }
     }
 
     /**